Changed all float types to double

+++ inc/geometry/matrix.h
@@ -10,7 +10,7 @@
 struct matrix {
         int row;
         int col;
-        float *matrix;
+        double *matrix;
 };
 
 /* CREATION */
@@ -39,10 +39,10 @@ struct matrix matrix_copy(const struct matrix m);
 struct matrix matrix_copy(const struct matrix m);
 
 /* Set the value of a row, col position, 0 if unable, 1 on success */
-int matrix_set(struct matrix m, int r, int c, float value);
+int matrix_set(struct matrix m, int r, int c, double value);
 
 /* Retrieve a value from the matrix, returns 0.0f on bad index */
-float matrix_get(const struct matrix m, int r, int c);
+double matrix_get(const struct matrix m, int r, int c);
 
 /* Check if two given matrices are equal, returns 1 if so, 0 otherwise */
 int matrix_equal(const struct matrix m, const struct matrix n);
@@ -65,23 +65,23 @@ const struct tuple transform(const struct tuple t, ...);
 struct matrix matrix_transpose(const struct matrix m);
 
 /* returns the determinant, matrix must be 2x2, returns 0 on fail */
-float matrix_2determinant(const struct matrix m);
+double matrix_2determinant(const struct matrix m);
 
 /* returns the determinant, matrix must be 3x3, returns 0 on fail */
-float matrix_3determinant(const struct matrix m);
+double matrix_3determinant(const struct matrix m);
 
 /* returns the minor of a 3x3 matrix at given point */
-float matrix_minor(struct matrix m, int row, int col);
+double matrix_minor(struct matrix m, int row, int col);
 
 /* return the given matrix with the specified column and row removed
  * returns the null matrix on failure */
 struct matrix submatrix(const struct matrix m, const int r, const int c);
 
 /* return the cofactor of a given 3x3 matrix at the given element, 0.0f on fail */
-float matrix_cofactor(const struct matrix m, const int r, const int c);
+double matrix_cofactor(const struct matrix m, const int r, const int c);
 
 /* return the determinant of a matrix, 0.0f if unable for some reason */
-float matrix_determinant(const struct matrix m);
+double matrix_determinant(const struct matrix m);
 
 /* test if a matrix is invertible, returns 0 if not, 1 if it is *
  * (a matrix is invertible if its determinant is non-zero
@@ -92,21 +92,21 @@ const int matrix_invertible(const struct matrix m);
 struct matrix matrix_inverse(const struct matrix m);
 
 /* return a matrix that will translate by x, y, z */
-struct matrix matrix_translate(const float x, const float y, const float z);
+struct matrix matrix_translate(const double x, const double y, const double z);
 
 /* return a matrix to scale a tuple */
-struct matrix matrix_scale(const float x, const float y, const float z);
+struct matrix matrix_scale(const double x, const double y, const double z);
 
 /* return a matrix that will rotate around the x-axis, radians */
-struct matrix matrix_rotate_x(const float radians);
+struct matrix matrix_rotate_x(const double radians);
 
 /* return a matrix that will rotate around the y-axis, radians */
-struct matrix matrix_rotate_y(const float radians);
+struct matrix matrix_rotate_y(const double radians);
 
 /* return a matrix that will rotate around the z-axis, radians */
-struct matrix matrix_rotate_z(const float radians);
+struct matrix matrix_rotate_z(const double radians);
 
 /* return a matrix that shears */
-struct matrix matrix_shear(const float xy, const float xz, const float yx, const float yz, const float zx, const float zy);
+struct matrix matrix_shear(const double xy, const double xz, const double yx, const double yz, const double zx, const double zy);
 
 #endif // __matrix_h__
